Fixed City

+++ PrismWeatherApp.Core/Models/City.cs
@@ -4,8 +4,8 @@
     {
         public int id { get; set; }
         public string name { get; set; }
-        public double latitude { get; set; }
-        public double longitude { get; set; }
+        public float latitude { get; set; }
+        public float longitude { get; set; }
         public string feature_code { get; set; }
         public string country_code { get; set; }
         public string country { get; set; }
